refactor: 패키지명 소문자화 + 복수형 처리
