+++ src/main/java/com/b1a4/cafeOn/configs/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package com.b1a4.cafeOn.Config;
+package com.b1a4.cafeOn.configs;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Info;
