+++ src/main/java/com/b1a4/cafeOn/configs/SecurityConfig.java
@@ -1,6 +1,8 @@
-package com.b1a4.cafeOn.Config;
+package com.b1a4.cafeOn.configs;
 
+import com.b1a4.cafeOn.security.JwtAuthenticationFilter;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -23,6 +25,8 @@ import static org.springframework.security.config.Customizer.withDefaults;
 @EnableWebSecurity  // Spring Security 활성화
 @Slf4j
 public class SecurityConfig {
+    @Autowired
+    private JwtAuthenticationFilter jwtAuthenticationFilter;
 
     @Bean   // Bean으로 등록했기때문에, 스프링이 PasswordEncoder 객체를 관리해서,
     // 다른곳에서 @Autowired PasswordEncoder passwordEncoder 선언 시, 스프링이 컨테이너 안의 이 Bean(BcryptPasswordEncoder)을 자동으로 찾아 주입함
@@ -41,7 +45,7 @@ public class SecurityConfig {
                         .permitAll()    // /, /api/auth/** 경로는 인증 안해도 되게 모두 허용하겠다!!(이코드 안쓰면 우리코드랑 관련없는 무슨 security 기본 로그인화면뜸)
                         .anyRequest().authenticated()); // 그 이외의 모든 경로는 인증 해야됨
 
-//        todo: filter 등록: 매 요청마다 CorsFilter를 실행한 후에 -> JwtAuthenticationFilter를 실행되게 순서 세팅
+//        filter 등록: 매 요청마다 CorsFilter를 실행한 후에 -> JwtAuthenticationFilter{}를 실행되게 순서 세팅
         http.addFilterAfter(jwtAuthenticationFilter, CorsFilter.class);
 
         return http.build();
@@ -63,5 +67,4 @@ public class SecurityConfig {
 
         return source;
     }
-
 }
