+++ src/main/java/com/b1a4/cafeOn/enums/UserProvider.java
@@ -1,3 +1,3 @@
-package com.b1a4.cafeOn.Enum;
+package com.b1a4.cafeOn.enums;
 
 public enum UserProvider { LOCAL, GOOGLE, KAKAO, NAVER }
