+++ src/main/java/com/b1a4/cafeOn/enums/UserRole.java
@@ -1,3 +1,3 @@
-package com.b1a4.cafeOn.Enum;
+package com.b1a4.cafeOn.enums;
 
 public enum UserRole { USER, ADMIN }
